Added Redis to project and implemented the FrontierQueue addURL() and getNext() methods
I installed Redis for Mac using homebrew and I found the Jedis jar file
to include in my Java project here:
https://mvnrepository.com/artifact/redis.clients/jedis/2.4.2

NOTE: A useful Java-specific Redis tutorial exists here:
https://www.tutorialspoint.com/redis/index.htm

+++ SearchEngine/src/webpage/URL.java
@@ -33,7 +33,7 @@ public class URL {
     private String subDomain;
     private String path;
     private String anchorText;
-    private byte priority;
+    private double priority;
     private String tag;
 
     public String getURL() {
@@ -108,11 +108,11 @@ public class URL {
         this.anchorText = anchorText;
     }
 
-    public byte getPriority() {
+    public double getPriority() {
         return priority;
     }
     
-    public void setPriority(byte priority) {
+    public void setPriority(double priority) {
         this.priority = priority;
     }
